% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmconf-package.R
\docType{package}
\name{unmconf}
\alias{unmconf}
\alias{package-unmconf}
\title{unmconf: Modeling with Unmeasured Confounding}
\description{
A package that fits and assesses Bayesian multi-staged regression models while accounting for unmeasured confounders.
}
