% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runm.R
\name{runm}
\alias{runm}
\title{Generate synthetic data}
\usage{
runm(
  n,
  response = "norm",
  confounder1 = "norm",
  confounder2 = NA,
  type = "int",
  missing_prop = 0.8
)
}
\arguments{
\item{n}{Number of observations}

\item{response}{\code{"norm"}, \code{"bin"}, \code{"pois"}, or \code{"gam"}}

\item{confounder1}{\code{"norm"} or \code{"bin"}}

\item{confounder2}{\code{"norm"} or \code{"bin"}}

\item{type}{\code{"int"} or \code{"ext"}}

\item{missing_prop}{Proportion of missing values. Only used when \code{type = "int"}}
}
\value{
A \code{tibble}
}
\description{
\code{\link[=runm]{runm()}} generates synthetic data for use in examples of models with
unmeasured confounders (can take up to two unmeasured confounders). Defaults to
one unmeasured confounder.
}
\examples{

# internally validated data
runm(20)
runm(20, missing_prop = .50)

runm(20, response = "bin")
runm(20, response = "pois")
runm(20, response = "gam")
runm(20, response = "norm", confounder1 = "bin")
runm(20, response =  "bin", confounder1 = "bin")
runm(20, response = "pois", confounder1 = "bin", missing_prop = .7)
runm(20, response = "gam", confounder1 = "bin", missing_prop = .7)

runm(20, response = "norm", confounder1 = "bin", confounder2 = "norm")
runm(20, response =  "bin", confounder1 = "bin", confounder2 = "norm")
runm(20, response = "pois", confounder1 = "bin", confounder2 = "bin", missing_prop = .7)
runm(20, response = "gam", confounder1 = "bin", confounder2 = "bin", missing_prop = .7)

runm(c(15,5), type = "ext")
runm(c(15,5), response = "norm", confounder1 = "norm", type = "ext")
runm(c(15,5), response = "norm", confounder1 = "bin", confounder2 = "norm", type = "ext")
}
