% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runm.R
\name{runm}
\alias{runm}
\alias{runm_full}
\alias{runm_int}
\alias{runm_ext}
\title{Generate synthetic data}
\usage{
runm(
  n,
  response,
  response_param = NULL,
  response_model_coefs,
  treatment_model_coefs,
  covariate_fam_list,
  covariate_param_list,
  unmeasured_fam_list,
  unmeasured_param_list,
  type = "int",
  missing_prop = 0.8
)

runm_full(
  n,
  response,
  response_param,
  response_model_coefs,
  treatment_model_coefs,
  covariate_fam_list,
  covariate_param_list,
  unmeasured_fam_list,
  unmeasured_param_list,
  treatment = TRUE
)

runm_int(
  n,
  response,
  response_param,
  response_model_coefs,
  treatment_model_coefs,
  covariate_fam_list,
  covariate_param_list,
  unmeasured_fam_list,
  unmeasured_param_list,
  missing_prop
)

runm_ext(
  n_main,
  n_external,
  response,
  response_param,
  response_model_coefs,
  treatment_model_coefs,
  covariate_fam_list,
  covariate_param_list,
  unmeasured_fam_list,
  unmeasured_param_list
)
}
\arguments{
\item{n}{Number of observations}

\item{response}{\code{"norm"}, \code{"bin"}, \code{"pois"}, or \code{"gam"}}

\item{response_param}{Nuisance parameters for response type.}

\item{response_model_coefs}{A named vector of coefficients to generate data
from the response model. This must include an intercept (\verb{"int" = }), a
coefficient for each covariate specified, a coefficient for each unmeasured
confounder, and a treatment coefficient (\verb{"x" = }).}

\item{treatment_model_coefs}{A named vector of coefficients to generate data
from the treatment model. This must include an intercept (\verb{"int" =}), a
coefficient for each covariate specified, and a coefficient for each
unmeasured confounder.}

\item{covariate_fam_list}{A list of either \code{"norm"} or \code{"bin"}, where the
length of the list matches the number of covariates in the model.}

\item{covariate_param_list}{A list of parameters for the respective
distributions in \code{covariate_fam_list}, where the length of the list matches
the length of \code{covariate_fam_list}.}

\item{unmeasured_fam_list}{A list of either \code{"norm"} or \code{"bin"}, where the
length of the list matches the number of unmeasured confounders in the
model. This can be at most a length of 2 to pair with \code{\link[=unm_glm]{unm_glm()}}.}

\item{unmeasured_param_list}{A list of parameters for the respective
distributions in \code{unmeasured_fam_list}, where the length of the list
matches the length of \code{unmeasured_fam_list}.}

\item{type}{\code{"int"} or \code{"ext"}}

\item{missing_prop}{Proportion of missing values. Only used when \code{type = "int"}}

\item{treatment}{Input documentation here.}

\item{n_main}{Input documentation here.}

\item{n_external}{Input documentation here.}
}
\value{
A \code{tibble}
}
\description{
\code{\link[=runm]{runm()}} generates customized synthetic data for use in examples of models
with unmeasured confounders. Currently set up to have at most 2 unmeasured
confounders to pair with \code{\link[=unm_glm]{unm_glm()}}.
}
\examples{

runm(
  n = 100,
  response = "norm",
  response_param = c("si_y" = 1),
  response_model_coefs = c("int" = -1, "z1" = .4, "z2" = .5, "z3" = .4,
                           "u1" = .75, "u2" = .75, "x" = .75),
  treatment_model_coefs = c("int" = -1, "z1" = .4, "z2" = .5, "z3" = .4,
                            "u1" = .75, "u2" = .75),
  covariate_fam_list = list("norm", "bin", "norm"),
  covariate_param_list = list(c(mean = 0, sd = 1), c(.3), c(0, 2)),
  unmeasured_fam_list = list("norm", "bin"),
  unmeasured_param_list = list(c(mean = 0, sd = 1), c(.3)),
  type = "int",
  missing_prop = .80
)
}
