% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unm_summary.R
\name{unm_summary}
\alias{unm_summary}
\title{Generate synthetic data}
\usage{
unm_summary(mod, data, quantiles = c(0.025, 0.975))
}
\arguments{
\item{mod}{Output from \code{\link[=unm_glm]{unm_glm()}}.}

\item{data}{The data \code{mod} was generated with.}

\item{quantiles}{A numeric vector of quantiles.}
}
\value{
A \code{tibble}
}
\description{
\code{\link[=unm_summary]{unm_summary()}} produces result summaries of the results from the model fitting
function, \code{\link[=unm_glm]{unm_glm()}}. The table of results are summarized from the MCMC draws of
the posterior distribution.
}
