% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmconf-package.R
\docType{package}
\name{unmconf}
\alias{unmconf}
\alias{unmconf-package}
\title{unmconf: Modeling with Unmeasured Confounding}
\description{
Tools for fitting and assessing Bayesian multi-staged regression models that
account for unmeasured confounders.
}
